Add files via upload
The primary goal of this survey was to collect information about victims that may or may not have experienced crime, and to estimate the amount of incidents and incident reports filed.  This was a prospective observational study where the respondents were asked household and crime questions within six months of the committed crime, which could included: sexual assault, robbery, assault with weapon, burglary, larceny, and motor vehicle theft.  This data was collected utilizing a stratified multistage cluster sample, as respondents were divided into subgroups by sex (gender) and age, then clustered into other subgroups like income, living type, education levels, race, sexual orientation, citizenship, and so on. As such, and considering how much data has been collected within this survey, the NCVS applies ‘weights’ to some parts of the data more than others, which would equalize and make the sampling procedure more unbiased. The overall sample seems to be unbiased as all persons in the United States 12 years of age and older were interviewed in each randomly selected household (which is nearly 150,000 households in a year) found through the census bureau.  An overall summary of this data shows that there are 8043 total observations across the total 81 columns. This study had the results of 1811 unique households at a given interview and 83 unique household records in a collection quarter. There are only 78 actual variables (not including the year, IDHH, and ICPSR columns). The variables can be split into household and crime specific variables. Common household variables focused on in this analysis include: sex, age, income, housing, education, vehicle, sexual orientation, living type, and marital status. The crime variables focused on in this analysis are: known offender, sexual assault, and vehicle theft.